requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.2.6
ollama==0.6.0
opencv-python==4.12.0.88
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.2.1
PyYAML==6.0.2
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.15.0
